Move filter function to util
